extract tests/core/test_evaluate_chat_response.py
