+++ tests/core/test_evaluate_functions.py
@@ -4,7 +4,6 @@ import itertools
 
 import pytest
 
-from flexeval.core.evaluate_chat_response import evaluate_chat_response
 from flexeval.core.evaluate_from_data import evaluate_from_data
 from flexeval.core.evaluate_generation import evaluate_generation
 from flexeval.core.evaluate_multiple_choice import evaluate_multiple_choice
@@ -16,7 +15,6 @@ from flexeval.core.metric import ExactMatch
 from flexeval.core.prompt_template import Jinja2PromptTemplate
 from flexeval.core.reward_model.pairwise_judge_reward_model import PairwiseJudgeRewardModel
 from tests.dummy_modules import (
-    DummyChatDataset,
     DummyGenerationDataset,
     DummyLanguageModel,
     DummyMultipleChoiceDataset,
@@ -25,47 +23,6 @@ from tests.dummy_modules import (
 )
 from tests.dummy_modules.reward_bench_dataset import DummyRewardBenchDataset
 from tests.dummy_modules.reward_lm import DummyRewardLanguageModel
-
-
-@pytest.mark.parametrize(
-    ("use_few_shot", "max_instances", "use_tools", "batch_size"),
-    list(itertools.product([True, False], [None, 1], [True, False], [1, 3])),
-)
-def test_evaluate_chat_response(use_few_shot: bool, max_instances: int, use_tools: bool, batch_size: int) -> None:
-    few_shot_generator = None
-    if use_few_shot:
-        few_shot_generator = RandomFewShotGenerator(dataset=DummyChatDataset(), num_shots=1, num_trials_to_avoid_leak=0)
-
-    metrics, outputs = evaluate_chat_response(
-        language_model=DummyLanguageModel(),
-        gen_kwargs={},
-        eval_dataset=DummyChatDataset(
-            use_tools=use_tools,
-        ),
-        few_shot_generator=few_shot_generator,
-        metrics=[],
-        batch_size=batch_size,
-        max_instances=max_instances,
-    )
-    assert isinstance(metrics, dict)
-    assert metrics["finish_reason_ratio-length"] == 1.0
-    assert isinstance(outputs, list)
-
-    if max_instances is not None:
-        assert len(outputs) <= max_instances
-
-    # If the system message in "messages", few-shot examples should be inserted after the system message.
-    # Therefore, in any case the system message should be in the first turn.
-    assert outputs[0]["extra_info"]["messages"][0]["role"] == "system"
-
-    if use_tools:
-        assert isinstance(outputs[0]["extra_info"]["tool_calls"], list)
-        assert isinstance(outputs[0]["extra_info"]["tools"], list)
-        assert metrics["tool_call_validation_result_ratio-CompleteToolCall"] == 1.0
-    else:
-        assert "tool_calls" not in outputs[0]["extra_info"]
-        assert "tools" not in outputs[0]["extra_info"]
-        assert metrics["tool_call_validation_result_ratio-TextOnly"] == 1.0
 
 
 @pytest.mark.parametrize("use_few_shot", [True, False])
